move api/sort to models/sort
